fix(ffi,build): codex native-review hardening
Address codex review findings on the FFI build (no ABI/behavior change; verified
by rebuild + an Android FFI-load run):

- ffi/tim2tox_ffi_exports.map + ffi/CMakeLists.txt: export ONLY the C FFI ABI
  (Dart* + tim2tox_ffi_*) from the Android .so via an ELF version-script,
  hiding the ~5k C++ mangled symbols and the vendored libsodium/toxcore symbols
  that previously leaked into the host Flutter process. Scoped to Android only —
  Linux keeps the full export surface for the native test harness (which dlsym's
  tox_* directly). Verified: the arm64 .so drops to 305 exports (Dart*/
  tim2tox_ffi_* only, 0 _Z), and the app still registers + connects to DHT, so
  both the binary-replacement (Dart*) and platform (tim2tox_ffi_*) paths load.

- source/CMakeLists.txt: scope -Wno-error=format to NOT APPLE. It only masks the
  benign LP64 uint64_t/int64_t -> %llu/%lld mismatch on Android/Linux; on Apple
  uint64_t IS unsigned long long, so -Werror=format stays on (the macOS FFI
  rebuilds clean). Also fix the one genuinely uncast site at
  V2TIMManagerImpl.cpp:2522 (uint64_t ts -> %llu).

+++ source/V2TIMManagerImpl.cpp
@@ -2519,7 +2519,7 @@ V2TIMString V2TIMManagerImpl::SendGroupPrivateTextMessage(const V2TIMString& gro
     }
     uint64_t ts = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
     char msg_id_buf[80];
-    snprintf(msg_id_buf, sizeof(msg_id_buf), "gp%llu-%llu", ts, (unsigned long long)msg_id);
+    snprintf(msg_id_buf, sizeof(msg_id_buf), "gp%llu-%llu", (unsigned long long)ts, (unsigned long long)msg_id);
     V2TIMString msgID = msg_id_buf;
     if (callback) {
         V2TIMMessage resultMsg;
